requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
openai==0.28.1
pdfplumber==0.10.3
google-generativeai==0.3.0
fpdf2==2.7.4
python-docx==1.1.0
streamlit